oh-tab-ycg2: extract text sanitizers
